SampleSetAssignment2/requirements.txt
streamlit==1.17.0
altair==4.2.2
sentence_transformers==3.1.0
transformers==4.44.2
faiss-cpu==1.8.0.post1
transformers==4.44.2
PyPDF2==1.26.0
